intrest rate is taken as double not int

+++ FVCalc.java
@@ -3,8 +3,8 @@ public class FVCalc {
 	public static void main(String[] args) {
 		// Replace this comment with your code
 		int currentValue = Integer.parseInt(args[0]);
-		int interstRate = Integer.parseInt(args[1]);
-		double percentageRate = (double) interstRate / 100;
+		double interstRate = Double.parseDouble(args[1]);
+		double percentageRate = interstRate / 100;
 		int years = Integer.parseInt(args[2]);
 
 		int futureValue = (int) (currentValue * Math.pow((1 + percentageRate), years));
